+++ src/Settings.php
@@ -3,7 +3,7 @@
 /**
  * Title: iDEAL Advanced v3 gateway settings
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
@@ -117,6 +117,7 @@ class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Settings extends Pronamic_WP_Pay_
 			'meta_key'    => '_pronamic_gateway_email',
 			'title'       => __( 'E-mail address', 'pronamic_ideal' ),
 			'tooltip'     => sprintf(
+				/* translators: %s: admin email */
 				__( 'E-mail address, e.g. %s', 'pronamic_ideal' ),
 				get_option( 'admin_email' )
 			),